+++ DOTSSimulator/Assets/Scripts/Boids/Energy/Producers/FoodSourceComponent.cs
@@ -5,11 +5,11 @@ namespace Simulator.Boids.Energy.Producers
 {
     public struct FoodSourceComponent : IComponentData
     {
-        public float EnergyLevel;
-        public float RegenarationRate;
-        public float MaxEnergyLevel;
+        public decimal EnergyLevel;
+        public decimal RegenarationRate;
+        public decimal MaxEnergyLevel;
         public float FeedingRadius;
-        public float EffectiveSize => EnergyLevel == 0 ? 0 : (EnergyLevel / 10f);
+        public float EffectiveSize => EnergyLevel == 0 ? 0 : ((float)EnergyLevel / 10f);
 
         public float EffectiveDistance(float3 position, float3 foodPosition)
         {
